Moved the PlaybookProcess out of CloudDeployProgress
Current PlaybookProcess is in CloudDeployProgress page. It is
used by SelectServersToProvision page. It should be a shared
component.
